Properly handle popupmenu sizes for popup position calculation.
Previously the popupmenu would always report its raw preferred size (i.e. without the maximum size constraints applied) to the position calculation algorithm. Due to this a popup, which would fit on the screen after it has been wrapped in the scrollable container would be moved. We'll now correctly report the effective size of the popup as the preferred/minimum/maximum size.