Menu: Safely check type of 'checkIconFactory' factory property
As the type can only mismatch when darklaf is being uninstalled we don't care what happens in this case. Simply don't crash. Moreover this only happens on windows where the window decorations may change the component hierarchy during uninstallation to return the menu bar to the root pane.
Currently we don't make use of this property at all, but to keep the implementation close to the swing internal one we keep the method for now.
Relates to #234