Decorations: Remove non jni install path
The client-property approach for java >= 12 resulted in window resizing failing
if it causes the window to be realized. Setting the 'fullSizeContentView' and
'transparentTitleBar' client-properties resulted in a native resize happening
inside 'Component#reshape' causing unexpected resize results.
We replace it with our jni approach to have better control over what happens to the window sizes.

Fixes #230