Restore previous Laf when creating the base laf
When we instantiate a Darklaf we have to create the base laf.
For macOS this may require us to install the system laf in the
UIManager. We should restore it the way we found it to ensure
no surprises if one instantiates Darklaf without trying to install
it directly afterwards. Because this makes the instantiation heavier
we skip the restoring if we know we are installing the laf directly
afterwards e.g. when installing through the LafManager.